juraj-main/Backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Speech-to-Text (Whisper)
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# NLP and RAG pipeline
langchain==0.0.350
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0
accelerate==0.25.0
llama-cpp-python==0.2.20

# Text-to-Speech
TTS==0.22.0
pyttsx3==2.90
soundfile==0.12.1

# Language detection
langdetect==1.0.9

# Utilities
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
pathlib==1.0.1

# Optional GPU support (uncomment if using CUDA)
# torch>=2.0.0+cu118
# torchaudio>=2.0.0+cu118

# Development and testing
pytest==7.4.3
httpx==0.25.2